// src/components/organisms/Header/Header.types.ts
// ============================================
// HEADER - Types & Interfaces
// ============================================

export interface HeaderProps {
  transparent?: boolean;
}

export interface DropdownItem {
  subtitle: string;
  href: string;
  image: ImageMetadata;
  description: string;
}